new throwing method implemented (still ne to be putted correctly)

+++ Brick breaker/Ball.cpp
@@ -71,7 +71,7 @@ void Ball::bounce(int angle) {
 	}
 }
 
-int Ball::getSpeed() {
+float Ball::getSpeed() {
 	return speed;
 }
